DOY-276 create コンポーネントをAboutCardWithImageLeft/Rightに分割
